Applied patch for Linux g++-4.2 and 64-bit
found patch at: http://sourceforge.net/tracker/?func=detail&aid=2792937&group_id=215539&atid=1034076

+++ asyncdns.cpp
@@ -101,7 +101,7 @@ C_AsyncDNS::~C_AsyncDNS()
 	return 0;
 }
 
-int C_AsyncDNS::resolve(char *hostname, unsigned long *addr)
+int C_AsyncDNS::resolve(char *hostname, uint32_t *addr)
 {
 	//return 0 on success, 1 on wait, -1 on unresolvable
 	int x;
